requirements.txt
discord.py
python-dotenv
requests

#pip install -r requirements.txt
